feat: added helpers, refactored
