+++ src/helper/parseJson.ts
@@ -1,0 +1,7 @@
+export function parseJson<T>(jsonString: string): T | null {
+  try {
+    return JSON.parse(jsonString) as T;
+  } catch (error) {
+    return null;
+  }
+}
