+++ src/helper/uuidValidation.ts
@@ -1,0 +1,10 @@
+import { parse as parseUUID } from 'uuid';
+
+export const isValidUUID = (uuid: string): boolean => {
+  try {
+    parseUUID(uuid);
+    return true;
+  } catch (error) {
+    return false;
+  }
+};
